refactor: 변경된 패키지 구조를 test 코드에 반영
